--- Sources/yaru/SubCommand.swift
@@ -1,17 +1,0 @@
-//
-//  SubCommand.swift
-//  
-//
-//  Created by Yuya Hirayama on 2022/03/10.
-//
-
-import Foundation
-
-enum SubCommand: String {
-    case `init` = "init"
-    case current = "current"
-    case push = "push"
-    case enqueue = "enq"
-    case next = "next"
-    case all = "all"
-}
